code organization works OK
