+++ package.json
@@ -9,10 +9,11 @@
   "author": "Sofia Neira, Camilo Murillo",
   "license": "ISC",
   "dependencies": {
-    "axios": "^1.8.4",
+    "axios": "^1.9.0",
     "cors": "^2.8.5",
     "express": "^5.1.0",
     "form-data": "^4.0.2",
+    "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.2"
   }
 }
