+++ routes/autorizar.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+const backendController = require('../controllers/backendController');
+
+// POST /autorizar
+router.post('/', backendController.autorizarUsuario);
+
+module.exports = router;
